Change Phenotype::mutate to mutate individual instead returning a new one
Returning a new individual on every call to mutate is very expensive for
large phenotypes. Especially, since not every call to mutate will likely
cause a mutation, as the mutation rate is often below 1. This makes the
copies just unnecessary. Phenotype requires the Copy trait, such that
the Simulator can simply copy the individual beforehand if a distinc
individual for the mutation is needed.